pyproject.toml
[tool.poetry]
name = "gitutor"
version = "0.5.2"
description = "Git wrapper that offers more user friendly commands"
readme = "README.md"
authors = ["AMAI"]
license = "MIT"
homepage = "https://gitutor.io"
keywords = ["git", "python", "cli"]
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
  { include = "gitutor"}
]

[tool.poetry.dependencies]
python = "^3.6"
click = "7.1.2"
GitPython = "3.1.3"
gitdb="4.0.5"
smmap="3.0.4"
pyinquirer="1.0.3"
pygithub="1.51"
colorama="0.4.3"
requests="^2.20.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
gt = 'gitutor.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
